Add the landing page and restructure project directories
Add the landing page for the front-end part of the application.
However, the routing of the landing page hasn't been fully implemented.
Future commit shall implement the routing using React Router,
and make the pages fully navigable.

This commit also restructure the project directories so that
the separation between client (front-end) and server (back-end) parts
of the application is made clear. They have their own package.json
and dependencies, and should be able to be deployed independently.