Add authentication and route protection
This commit enables user to log in and log out from the front-end
of the application by requesting API call that is consumed by the back-
end server. The client will also check if the session cookie is still
valid in a 'stateful' manner that is persisted throughout the session.

Protected routes were also implemented in this commit, prompting user to
log in if they try to access the protected routes. This is done by
checking if there is a valid session associated with the session cookie
in the client by API call made to the back-end.

This whole feature is done by implementing the underlying 'AuthContext'.
This context provides for the whole application, which is easily
usable by using the 'useAuth' hook.