Implement goal tracking and nutrition entry
This commit enables user to track their current goal and target
nutrition to reach each day. To assist this endeavour, user can now
enter their nutrition intake entry on the goal page. User can also
see the overview of their goal progress on the goal page; choosing a
date on the calendar to see their entries for that particular day.

The CRUD operations for the nutrition entry have been tested and it
worked well at the time of this commit. The timezone used for the
operations is based on local timezone of the user computer. It works
fine with GMT+10 timezone as per user journey testing. However, other
timezones haven't been tested.

Next commit should refine the UI for the user to alter their goal
and target nutrition based on the selected goal. The backend APIs for
this purpose are still rudimentary, therefore there might be breaking
changes as the process ongoing.